IT gives kaprekar between between two given numbers
Def:In mathematics, a Kaprekar number for a given base is a non-negative integer, the representation of whose square in that base can be split into two parts that add up to the original number again. For instance, 45 is a Kaprekar number, because 45² = 2025 and 20+25 = 45.

So, if we give input as 1 & 10 then the answer will be 1 & 9.